+++ XBOOK/XBOOK.Web/Controllers/JournalEntryController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using XBOOK.Common.Helpers;
 using XBOOK.Data.Model;
 using XBOOK.Service.Interfaces;
 
@@ -9,9 +11,12 @@ namespace XBOOK.Web.Controllers
     public class JournalEntryController : BaseAPIController
     {
         private readonly IJournalEntryService _journalEntryService;
-        public JournalEntryController(IJournalEntryService journalEntryService)
+        private readonly IAuthorizationService _authorizationService;
+
+        public JournalEntryController(IJournalEntryService journalEntryService, IAuthorizationService authorizationService)
         {
             _journalEntryService = journalEntryService;
+            _authorizationService = authorizationService;
         }
 
         [HttpPost("[action]")]
@@ -28,6 +33,9 @@ namespace XBOOK.Web.Controllers
         [HttpPost("[action]")]
         public async Task<IActionResult> CreateJournalEntry(JournalEntryModelCreate request)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, "Journal Entries", Operations.Create);
+            if (!result.Succeeded)
+                return Unauthorized();
             return Ok(await _journalEntryService.CreateJournalEntry(request));
         }
 
@@ -40,12 +48,18 @@ namespace XBOOK.Web.Controllers
         [HttpPut("[action]")]
         public async Task<IActionResult> UpdateJournalEntry(JournalEntryModelCreate request)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, "Journal Entries", Operations.Update);
+            if (!result.Succeeded)
+                return Unauthorized();
             return Ok(await _journalEntryService.UpdateJournalEntry(request));
         }
 
         [HttpPost("[action]")]
         public async Task<IActionResult> DeleteJournal([FromBody]List<Deleted> request)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, "Journal Entries", Operations.Delete);
+            if (!result.Succeeded)
+                return Unauthorized();
             return Ok(await _journalEntryService.DeleteJournalEntry(request));
         }
     }
